[Fix] #166 - 표지 캐시 url 변경·실패 경로 보정, 캐시 클래스 파일 분리
